finish

+++ get_next_line/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: gafreire <gafreire@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/19 17:19:28 by gafreire          #+#    #+#             */
-/*   Updated: 2025/02/01 17:07:25 by gafreire         ###   ########.fr       */
+/*   Updated: 2025/03/09 20:31:32 by gafreire         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,6 +29,6 @@ char	*extract_remaining(char *line_buffer);
 char	*ft_strjoin(char *s1, char *s2);
 char	*ft_strchr(char *s, int c);
 char	*ft_strdup(char *s1);
-size_t	ft_strlen(char *s);
+int	    ft_strlen(char *s);
 
 #endif
